Improved Minecraft API * added implementations of block meta

+++ internalpluginmanager-minecraft-api/src/main/java/com/blockhaus2000/ipm/minecraft/block/meta/BrewingStandBlockMeta.java
@@ -17,9 +17,8 @@
  */
 package com.blockhaus2000.ipm.minecraft.block.meta;
 
-
 public interface BrewingStandBlockMeta extends BlockMeta, Container {
-    int getRemainingBrewingTime();
+    long getRemainingBrewingTime();
 
     boolean isBrewing();
 }
